Render checkpoint run_metadata via shared JsonTree
Extract the styled JsonView wrapper from JsonContent into a reusable
JsonTree component, then render checkpoint.metadata.run_metadata in the
checkpoint detail tab. Also drops the unused custom JsonViewer.
