+++ src/modules/executions/ui/traces/JsonContent.tsx
@@ -1,6 +1,4 @@
 import type { ReactNode } from "react";
-import { JsonView, allExpanded, defaultStyles } from "react-json-view-lite";
-import "react-json-view-lite/dist/index.css";
 
 import { CodeBlock } from "@/shared/ui/CodeBlock";
 import {
@@ -8,6 +6,7 @@ import {
 	normalizeJsonVisualization,
 	selectPromotedMarkdownField,
 } from "../../domain/content-parser";
+import { JsonTree } from "./JsonTree";
 import { RenderedMarkdown } from "./MarkdownContent";
 import { ViewerFrame } from "./ViewerFrame";
 
@@ -21,38 +20,6 @@ type JsonRenderPlan = {
 	sizeLabel: string;
 	decodedFromJson: boolean;
 };
-
-const jsonViewStyles = {
-	...defaultStyles,
-	container: "font-mono text-xs leading-relaxed",
-	basicChildStyle: "my-0",
-	label: "text-foreground mr-1",
-	clickableLabel: "text-foreground mr-1 cursor-pointer",
-	nullValue: "text-muted-foreground italic",
-	undefinedValue: "text-muted-foreground italic",
-	numberValue: "text-primary",
-	stringValue: "text-success whitespace-pre-wrap",
-	booleanValue: "text-primary",
-	otherValue: "text-foreground",
-	punctuation: `${defaultStyles.punctuation} !text-muted-foreground`,
-	expandIcon: `${defaultStyles.expandIcon} !text-muted-foreground`,
-	collapseIcon: `${defaultStyles.collapseIcon} !text-muted-foreground`,
-	childFieldsContainer: "border-l border-border ml-2 pl-2",
-	quotesForFieldNames: true,
-};
-
-function JsonTree({ data }: { data: object }) {
-	return (
-		<div className="overflow-x-auto p-4">
-			<JsonView
-				data={data}
-				shouldExpandNode={allExpanded}
-				clickToExpandNode
-				style={jsonViewStyles}
-			/>
-		</div>
-	);
-}
 
 function stringifyJsonValue(value: unknown): string {
 	return JSON.stringify(value, null, 2) ?? String(value);
@@ -105,7 +72,7 @@ function PromotedMarkdownObject({
 						Metadata ({metadataEntries.length}{" "}
 						{metadataEntries.length === 1 ? "field" : "fields"})
 					</summary>
-					<div className="border-border bg-muted/20 mt-3 overflow-hidden rounded-md border">
+					<div className="border-border bg-muted/20 mt-3 overflow-hidden rounded-md border p-4">
 						<JsonTree data={metadata} />
 					</div>
 				</details>
@@ -163,7 +130,11 @@ function planParsedJson(
 
 	const copyText = stringifyJsonValue(value);
 	return {
-		rendered: <JsonTree data={value as object} />,
+		rendered: (
+			<div className="p-4">
+				<JsonTree data={value as object} />
+			</div>
+		),
 		copyText,
 		sizeLabel: `${copyText.length} chars decoded`,
 		decodedFromJson: wasStringEnvelope,
